// src/type/onboard.type.ts
export type TDataFollowOnboard = {
    username: string;
    firstName: string;
    lastName: string;
    photoUrl: string;
    followers: number;
    winRate: number;
    profit: number;
    isFollowing: boolean;
    isPremium: boolean;
    _id: number;
    lastFollower: number;
    profit7d: number;
    winRate7d: number;
};